+++ example/lib/page_visualizer.dart
@@ -42,7 +42,7 @@ class _PageVisualizerState extends State<PageVisualizer> {
   final ValueNotifier<TextureType> textureType =
       ValueNotifier(TextureType.fft2D);
   final ValueNotifier<double> fftSmoothing = ValueNotifier(0.8);
-  final ValueNotifier<bool> isVisualizerForPlayer = ValueNotifier(false);
+  final ValueNotifier<bool> isVisualizerForPlayer = ValueNotifier(true);
   final ValueNotifier<bool> isVisualizerEnabled = ValueNotifier(true);
   final ValueNotifier<RangeValues> fftImageRange =
       ValueNotifier(const RangeValues(0, 255));
@@ -52,12 +52,32 @@ class _PageVisualizerState extends State<PageVisualizer> {
   Timer? timer;
   SoundProps? currentSound;
   FftController visualizerController = FftController();
+  bool canBuild = false;
 
   @override
-  void dispose() {
-    SoLoud.instance.shutdown();
-    SoLoudCapture.instance.stopCapture();
-    super.dispose();
+  void initState() {
+    super.initState();
+
+    /// Initialize the player
+    SoLoud.instance.initialize().then((value) {
+      if (value == PlayerErrors.multipleInitialization) {
+        SoLoud.instance.disposeAllSound();
+      }
+      if (value == PlayerErrors.noError ||
+          value == PlayerErrors.multipleInitialization) {
+        _log.info('player started');
+        SoLoud.instance.setVisualizationEnabled(true);
+        SoLoud.instance.setGlobalVolume(1);
+        if (context.mounted) {
+          setState(() {
+            canBuild = true;
+          });
+        }
+      } else {
+        _log.severe('player starting error: $value');
+        return;
+      }
+    });
   }
 
   @override